+++ src/business/IBoVeiculo.java
@@ -7,6 +7,6 @@ import excecoes.BoException;
 
 public interface IBoVeiculo {
 	
-	public int totalVeiculoDisponivel(Filial filial, CategoriaVeiculo categoria)throws BoException ;
-	public int totalManutencoesPententes(Veiculo veiculo) throws BoException;
+	public long totalVeiculoDisponivel(Filial filial, CategoriaVeiculo categoria)throws BoException ;
+	public long totalManutencoesPententes(Veiculo veiculo) throws BoException;
 }
