+++ src/dao/DaoVeiculo.java
@@ -15,14 +15,14 @@ public class DaoVeiculo extends Dao<Veiculo> implements IDaoVeiculo {
 	}
 
 	@Override
-	public int totalVeiculoDisponivel(Filial filial, CategoriaVeiculo categoria)throws DaoException {
-		int total = 0;
+	public long totalVeiculoDisponivel(Filial filial, CategoriaVeiculo categoria)throws DaoException {
+		long total = 0;
 		try {
 			em = ConnectionFactory.getConnection();
 			Query query = em.createNamedQuery(TOTAL_DISPONIVEL);
 			query.setParameter("filial",filial);
 			query.setParameter("categoria", categoria);
-			total =(Integer) query.getSingleResult();
+			total =(Long) query.getSingleResult();
 		}catch (Exception e) {
 			e.printStackTrace();
 			throw new DaoException("ERRO AO BUSCAR TOTAL DE VEICULOS DISPONIVEIS POR FILIAL E CATEGORIA DE VEICULO ");
@@ -31,13 +31,13 @@ public class DaoVeiculo extends Dao<Veiculo> implements IDaoVeiculo {
 	}
 
 	@Override
-	public int totalManutencaoPendente(Veiculo veiculo) throws DaoException {
-		int total = 0;
+	public long totalManutencaoPendente(Veiculo veiculo) throws DaoException {
+		long total = 0;
 		try {
 			em = ConnectionFactory.getConnection();
 			Query query = em.createNamedQuery(TOTAL_MANUTENCAO_PENDENTE);
 			query.setParameter("veiculo",veiculo);
-			total =(Integer) query.getSingleResult();
+			total =(Long) query.getSingleResult();
 		}catch (Exception e) {
 			e.printStackTrace();
 			throw new DaoException("ERRO AO BUSCAR TOTAL DE MANUTENCÕES DO VEICULO");
